extracting storage vos from config package
